refactor array constructors into construction.py
